* gcov-io.h (GCOV_TAG_SAME_VALUE_HISTOGRAMS): New section. 	(gcov_read_counter): Add parameter. 	(merger_function, same_value_histograms_merger, profile_merger_for_tag): 	New. 	(gcov_read_summary): Add parameter to gcov_read_counter call. 	* gcov.c (read_count_file): Ditto. 	* libgcc.c (gcov_exit): Support profile merging hooks. 	* profile.c (same_value_histograms_label): New. 	(get_exec_counts): Add parameter to gcov_read_counter call. 	(get_histogram_counts): Support profile merging hooks. 	(instrument_values, compute_value_histograms): Place same value 	histograms into own section. 	(branch_prob): Call allocate_reg_info after find_values_to_profile. 	(init_branch_prob): Create same_value_histograms_label. 	(label_for_tag): Return the label. 	* profile.h (MAX_COUNTER_SECTIONS): Increase.
	* loop.c (loop_optimize): Avoid warning.
	* regclass.c (cannot_change_mode_set_regs): Ditto.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/hammer-3_3-branch@63251 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/loop.c
@@ -498,7 +498,7 @@ loop_optimize (f, dumpfile, flags)
 
   /* Allocate and initialize auxiliary loop information.  */
   loops_info = xcalloc (loops->num, sizeof (struct loop_info));
-  for (i = 0; i < loops->num; i++)
+  for (i = 0; i < (int) loops->num; i++)
     loops->array[i].aux = loops_info + i;
 
   /* Now find all register lifetimes.  This must be done after
